Time: 1 ms (74.07%), Space: 2.2 MB (72.49%) - LeetHub

+++ 0033-search-in-rotated-sorted-array/0033-search-in-rotated-sorted-array.rs
@@ -1,8 +1,8 @@
 impl Solution {
     pub fn search(nums: Vec<i32>, target: i32) -> i32 {
         let mut low = 0;
-        let mut high = nums.len() as i32 - 1;
-        
+        let mut high = nums.len()  -1;
+       
         while low <= high {
             let mid = (low + high) / 2;
             
@@ -27,5 +27,6 @@ impl Solution {
         }
         
         -1
+        
     }
 }
